+++ tools/test_ui_contract.py
@@ -25,9 +25,9 @@ if missing:errors.append("IDs usadas por JS inexistentes: "+", ".join(missing))
 required={
 "heroYear","heroEra","yearDisplay","yearMagnitude","yearEra","goYearBtn",
 "occurrenceCount","subjectCount","placeCount","eventCount",
-"searchInput","subjectTypeFilter","evidenceFilter","occurrenceTypeFilter","labelMode",
+"searchInput","subjectTypeFilter","evidenceFilter","occurrenceTypeFilter","certaintyFilter","precisionFilter","spatialFilter","labelMode",
 "resetFiltersBtn","worldMap","basemapLayer","occurrenceLayer","occurrenceList",
-"categorySummary","eventList","filterDrawer","layersDrawer","detailDrawer","subjectDetail","occurrenceDetail","contextLayer","developmentLayer","layerGastronomy","layerContexts","layerDevelopments","layerSafety","mapCoverageStatus","historyDrawer","historyTitle","historySubtitle","historySummary","historyTimeline","temporalNavigator","temporalRail","temporalDensity","temporalVerified","temporalEvents","temporalDevelopments","temporalCursor","temporalFocus","prevTemporalHitBtn","nextTemporalHitBtn","temporalMinLabel","temporalNavigatorHint","temporalMaxLabel"
+"categorySummary","evidenceLensSummary","evidenceLensFilterBtn","eventList","filterDrawer","layersDrawer","detailDrawer","subjectDetail","occurrenceDetail","evidenceReadingSection","evidenceReadingDetail","contextLayer","developmentLayer","layerGastronomy","layerContexts","layerDevelopments","layerSafety","mapCoverageStatus","unmappedRecordsPanel","historyDrawer","historyTitle","historySubtitle","historySummary","historyTimeline","temporalNavigator","temporalRail","temporalDensity","temporalVerified","temporalEvents","temporalDevelopments","temporalCursor","temporalFocus","prevTemporalHitBtn","nextTemporalHitBtn","temporalMinLabel","temporalNavigatorHint","temporalMaxLabel"
 }
 mr=sorted(required-set(p.ids))
 if mr:errors.append("Faltan IDs requeridas: "+", ".join(mr))